raw type optimization

+++ src/com/tonybeltramelli/desktop/core/scoring/LanguageBasedScoring.scala
@@ -4,7 +4,7 @@ import com.tonybeltramelli.desktop.util.Helper
 
 class LanguageBasedScoring extends AScoring
 {  
-	override def getScore(f: (Map[String, Double], Double), query: List[String]) : Double =
+	override def getScore(f: (Map[String, Int], Int), query: List[String]) : Double =
 	{
 	  val lambda : Double = 0.1 //0.1 for title queries and 0.7 for long queries, Jelinek-Mercer smoothing method 
 	  //according to Chengxiang Zhai and John Lafferty, A Study of Smoothing Methods for Language Models Applied to Information Retrieval. (2004)
